Seperate request handling
